unify spec worker update_weights via draft_runners
